requirements.txt
fastapi
uvicorn
pydantic[email]
python-box
pyYAML
ensure
joblib
types-PyYAML
python-multipart
fastapi-sqlalchemy
-e .